KD-4344: Borrowers/status security updates
This commit does three things:

1) Make borrowers/status update and follow failed login attempts
and block access to the endpoint when the maximum amount is
exceeded. This will prevent brute brute-forcing the user
accounts via this endpoint.

2) Block access to the endpoint if the card used for logging in
is reported lost. The card is likely to be in the wrong hands,
so we should say nothing about the patron, even with seemingly
successful login.

Additionally 'gonenoaddress' will no longer be considered same
as card lost. That just means that the patrons address is
unknown or in doubt, it does not imply that the card might be
in the wrong hands. We do not want the patrons to be blocked
out of e-services because of that.

3) Prevent the endpoint from leaking information about user
accounts (i.e. valid username but wrong password) in error
messages by just returning a generalized error message on
failed logins.